position options in the dropdown
